Unit tests for vehicle update, view, delete, task 44
